Create Present Date.bat
Click on it to create a Folder with present Date. If the folder is present, it does nothing. 
Copy and Paste the .bat file in the folder, where you want to create the new folders with just a click. No need to change the path.

Just copy and paste the whole bat file, not the shortcut.